% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_data}
\alias{df_data}
\title{Filtered result data (dataframe)}
\format{
A dataframe with 15 columns:
 \describe{
   \item{Site_ID}{Site ID}
   \item{Site_Name}{Site name}
   \item{Watereshed}{Watershed}
   \item{Date}{Date}
   \item{Year}{Year}
   \item{Parameter}{Parameter}
   \item{Result}{Result value}
   \item{Unit}{Result unit}
   \item{Depth}{Depth category}
   \item{Min}{Minimum acceptable value}
   \item{Max}{Maximum acceptable value}
   \item{Excellent}{Threshold between excellent and good data}
   \item{Best}{Whether excellent data is higher or lower than fair data}
   \item{Month}{Month}
   \item{Description}{Popup text for interactive graph}
 }
}
\usage{
df_data
}
\description{
Dataframe containing filtered result data.
}
\keyword{datasets}
