% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_data_all}
\alias{df_data_all}
\title{Result data (dataframe)}
\format{
A dataframe with at least 13 columns:
 \describe{
   \item{Site_ID}{Site ID}
   \item{Activity_Type}{Activity type}
   \item{Date}{Date}
   \item{Depth}{Sample depth}
   \item{Depth_Unit}{Depth unit}
   \item{Depth_Category}{Depth category}
   \item{Parameter}{Parameter}
   \item{Result}{Result value}
   \item{Result_Unit}{Result unit}
   \item{Lower_Detection_Limit}{Lower detection limit}
   \item{Upper_Detection_Limit}{Upper detection limit}
   \item{Detection_Limit_Unit}{Detection limit unit}
   \item{Qualifier}{Qualifier code}
 }
}
\usage{
df_data_all
}
\description{
Dataframe containing raw result data.
}
\keyword{datasets}
