% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_sites}
\alias{df_sites}
\title{Filtered site metadata (dataframe)}
\format{
A dataframe with 5-7 columns:
 \describe{
   \item{Site_ID}{Site ID}
   \item{Site_Name}{Site name}
   \item{Latitude}{Latitude}
   \item{Longitude}{Longitude}
   \item{Town}{Town}
   \item{State}{State}
   \item{Watershed}{Watershed name}
   \item{Group}{Group. Values: Warmwater, Coldwater, Saltwater, NA}
 }
}
\usage{
df_sites
}
\description{
Dataframe containing site metadata.
}
\keyword{datasets}
