% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_sites_all}
\alias{df_sites_all}
\title{Site metadata (dataframe)}
\format{
A dataframe with at least 8 columns:
 \describe{
   \item{Site_ID}{Site ID}
   \item{Site_Name}{Site name}
   \item{Latitude}{Latitude}
   \item{Longitude}{Longitude}
   \item{Town}{Town}
   \item{State}{State}
   \item{Watershed}{Watershed name}
   \item{Group}{Group. Values: Warmwater, Coldwater, Saltwater, NA}
   \item{Max_Surface_Depth_m}{Maximum surface depth, meters. Default 1.}
   \item{Max_Midwater_Depth_m}{Maximum midwater depth, meters}
   \item{Max_Depth_m}{Site depth, meters}
   \item{Location_Type}{Location type}
 }
}
\usage{
df_sites_all
}
\description{
Dataframe containing raw site metadata.
}
\keyword{datasets}
