% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WQdashboard-package.R
\docType{package}
\name{WQdashboard-package}
\alias{WQdashboard}
\alias{WQdashboard-package}
\title{WQdashboard: Water Quality Dashboard}
\description{
A water quality dashboard template.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NBEP/WQdashboard}
  \item Report bugs at \url{https://github.com/NBEP/WQdashboard/issues}
}

}
\author{
\strong{Maintainer}: Mariel Sorlien \email{msorlien@nbep.org}

}
\keyword{internal}
